// prisma/generated/type-graphql/resolvers/inputs/Directus_usersWhereInput.ts
export class Directus_usersWhereInput {
  @TypeGraphQL.Field(_type => [Directus_usersWhereInput], {
    nullable: true
  })
  AND?: Directus_usersWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [Directus_usersWhereInput], {
    nullable: true
  })
  OR?: Directus_usersWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [Directus_usersWhereInput], {
    nullable: true
  })
  NOT?: Directus_usersWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  id?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  first_name?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  last_name?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  email?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  password?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  location?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  title?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  description?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  tags?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  avatar?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  language?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  theme?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  tfa_secret?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  status?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  role?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  token?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  last_access?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  last_page?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  provider?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  external_identifier?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  auth_data?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => BoolNullableFilter, {
    nullable: true
  })
  email_notifications?: BoolNullableFilter | undefined;

  @TypeGraphQL.Field(_type => Directus_dashboardsListRelationFilter, {
    nullable: true
  })
  directus_dashboards?: Directus_dashboardsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Directus_filesListRelationFilter, {
    nullable: true
  })
  directus_files_directus_files_modified_byTodirectus_users?: Directus_filesListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Directus_filesListRelationFilter, {
    nullable: true
  })
  directus_files_directus_files_uploaded_byTodirectus_users?: Directus_filesListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Directus_flowsListRelationFilter, {
    nullable: true
  })
  directus_flows?: Directus_flowsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Directus_rolesNullableRelationFilter, {
    nullable: true
  })
  directus_roles?: Directus_rolesNullableRelationFilter | undefined;
}